// src/app/modulos/medico/modelos/paciente.interface.ts
export interface IPaciente{
    id: any;
    identificacion: any;
    primer_nombre: any;
    segundo_nombre: any;
    primer_apellido: any;
    segundo_apellido: any;
    edad: any;
    genero: any;
    fecha_nacimiento: any;
    telefono: any;
    correo: any;
    ocupacion: any;
    medico_id:any
}